requirements.txt
google-api-python-client
google-auth-oauthlib 
google-auth-httplib2
schedule
ffmpeg-python
yt-dlp
unidecode
flask
loguru
underthesea
soundfile
dm-haiku
openai-whisper